requirements.txt
feedparser
requests
groq
anthropic
httpx
readability-lxml
lxml_html_clean
beautifulsoup4
jinja2
